% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow.R
\name{checkExperiment}
\alias{checkExperiment}
\title{Coleta os experimentos do MLFlow e verifica se o nome do experimento
enviado por parametro já existe.}
\usage{
checkExperiment(experiment_name, client)
}
\arguments{
\item{experiment_name}{nome unico do experimento MLFlow}

\item{client}{objeto contendo o MLFlow client}
}
\value{
id do experimento do MLFlow ou 0 caso nao exista
}
\description{
Coleta os experimentos do MLFlow e verifica se o nome do experimento
enviado por parametro já existe.
}
