% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow.R
\name{finish_mlflow}
\alias{finish_mlflow}
\title{Sinaliza MLFlow de que o processamento acabou. Opcionalmente pode ser
utilizado também pra salvar metricas e parametros.}
\usage{
finish_mlflow(run_obj, df_parameter = NULL, df_metric = NULL, final_status)
}
\arguments{
\item{run_obj}{id da rodada recebido no retorno da funcao start_mlflow}

\item{df_parameter}{DataFrame de tags contendo as colunas: params e values}

\item{df_metric}{DataFrame de metricas contendo: params e values}

\item{final_status}{string com status final do MLFlow. Use MLFlowStatusEnum}
}
\description{
Sinaliza MLFlow de que o processamento acabou. Opcionalmente pode ser
utilizado também pra salvar metricas e parametros.
}
