% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Carrega dados de um dataset}
\usage{
load_data(infile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Esta funcao carrega um dataset mapeados no Catalogo de Dados da plataforma
UniSoma e retorna para a funcao R na forma de uma estrutura de dados
apropriada.
}
