% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow.R
\name{MLFlowStatusEnum}
\alias{MLFlowStatusEnum}
\title{Funcao que emula um enumerador com os valores esperados pelo MLFlow para
definicao dos status finais de uma rodada. Utilize o exemplo abaixo como
um dos argumentos da funcao finish_mlflow caso queira definir que a rodada
finalizou com sucesso:
MLFlowStatusEnum()$FINISHED}
\usage{
MLFlowStatusEnum()
}
\value{
string contendo um dos status esperados pelo MLFlow
}
\description{
Funcao que emula um enumerador com os valores esperados pelo MLFlow para
definicao dos status finais de uma rodada. Utilize o exemplo abaixo como
um dos argumentos da funcao finish_mlflow caso queira definir que a rodada
finalizou com sucesso:
MLFlowStatusEnum()$FINISHED
}
