% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow.R
\name{save_artifact}
\alias{save_artifact}
\title{Salva artefato no repositorio do MLFLow}
\usage{
save_artifact(run_obj, df_artifact = NULL, file_path_artifact = NULL)
}
\arguments{
\item{df_artfact}{DataFrame com dados a serem salvos como um artefato}

\item{file_path_artfact}{caminho do arquivo de artefato a ser salvo}
}
\description{
Esta funcao envia para o MLFlow um artefato para o MLFlow relacionado
a uma rodada. Pode ser enviado um data frame ou mesmo o caminho de um
arquivo físico. No caso de ser um data frame, ele será convertido em CSV
}
