% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow.R
\name{start_mlflow}
\alias{start_mlflow}
\title{Inicia publicacao no MLFlow}
\usage{
start_mlflow(experiment_name)
}
\arguments{
\item{experiment_name}{nome unico do experimento MLFlow}
}
\value{
id da rodada necessario para finalizacao da publicacao no MLFlow
}
\description{
Tipicamente esta funcao deve ser posicionada no inicio do script R para
que o MLFlow possa medir o tempo de execução do modelo.
}
