% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Carrega dados de um dataset}
\usage{
load_data(entity_catalog)
}
\arguments{
\item{entity_catalog}{id of data catalog entity}
}
\value{
dados coletada da base fonte segundo o catalogo
}
\description{
Esta funcao carrega um dataset mapeados no Catalogo de Dados da plataforma
UniSoma e retorna para a funcao R na forma de uma estrutura de dados
apropriada.
}
